게시글 인가 기능 구현

+++ src/controllers/article.controller.js
@@ -1,4 +1,3 @@
-
 const articleService = require('../services/article.service');
 
 exports.getAllArticles = async (req, res, next) => {
@@ -21,27 +20,49 @@ exports.getArticleById = async (req, res, next) => {
 
 exports.createArticle = async (req, res, next) => {
     try {
-        const result = await articleService.createArticle(req.body);
-        res.status(201).json(result);
-    } catch (err) {
-        next(err);
+        const userId = req.user.id;
+        const result = await articleService.createArticle({ ...req.body, userId });
+        console.log(`✅ 게시글 등록 완료:`, result);
+        res.status(201).json({ message: 'Successfully created', result });
+    } catch (error) {
+        console.error(`error:`, error);
+        return next(err);
     }
 };
 
 exports.updateArticle = async (req, res, next) => {
     try {
+        const userId = req.user.id;
+        const article = await articleService.getArticleById(req.params.id);
+
+        console.log('로그인한 유저 ID:', userId);
+        console.log('게시글 작성자 ID:', article.userId);
+
+        if (Number(article.userId) !== Number(userId)) {
+            return res.status(403).json({ message: 'Unauthorized' });
+        }
+
         const result = await articleService.updateArticle(req.params.id, req.body);
-        res.status(200).json(result);
+        console.log(`✅ 업데이트 완료:`, result);
+        res.status(200).json({ message: 'Successfully updated', result });
     } catch (err) {
-        next(err);
+        return next(err);
     }
 };
 
 exports.deleteArticle = async (req, res, next) => {
     try {
+        const userId = req.user.id;
+        const article = await articleService.getArticleById(req.params.id);
+
+        if (Number(article.userId) !== Number(userId)) {
+            return res.status(403).json({ message: 'Unauthorized' });
+        }
+
         const result = await articleService.deleteArticle(req.params.id);
-        res.status(200).json(result);
+        console.log(`✅ 삭제 완료:`, result);
+        res.status(200).json({ message: 'Successfully deleted' });
     } catch (err) {
-        next(err);
+        return next(err);
     }
 };
